Hehe

+++ src/main/java/sypztep/crital/common/data/CritOverhaulConfig.java
@@ -11,50 +11,47 @@ import java.nio.file.Paths;
 import java.util.List;
 
 public class CritOverhaulConfig {
-    private static final Gson gson = new GsonBuilder().setPrettyPrinting().create();
-    private static final String CONFIG_FILE_PATH = FabricLoader.getInstance().getConfigDir().resolve("crital/crit_chances.json").toString();
+    public static final Gson gson = new GsonBuilder().setPrettyPrinting().create();
+    private static final String CONFIG_FILE_PATH = FabricLoader.getInstance().getConfigDir().resolve("newcritoverhaul/crit_chances.json").toString();
     private final CritOverhaulData critOverhaulData;
-
+    private static final CritOverhaulData CLIENT_DEFAULT_DATA = new CritOverhaulData();
+    public static String getConfigFilePath() {
+        return CONFIG_FILE_PATH;
+    }
     public CritOverhaulConfig() {
         this.critOverhaulData = loadConfig();
     }
-
-
-    private CritOverhaulData loadConfig() {
+    public CritOverhaulData loadConfig() {
         try {
             Path configFilePath = Paths.get(CONFIG_FILE_PATH);
-
             if (Files.exists(configFilePath)) {
                 try (Reader reader = new FileReader(configFilePath.toFile())) {
                     return gson.fromJson(reader, CritOverhaulData.class);
                 }
             } else {
-                // If the config file doesn't exist, create a default one
-                CritOverhaulData defaultData = new CritOverhaulData();
-                saveConfig(defaultData);
-                return defaultData;
+                saveConfig(CLIENT_DEFAULT_DATA);
+                return CLIENT_DEFAULT_DATA;
             }
         } catch (IOException e) {
             e.printStackTrace();
-            return new CritOverhaulData(); // Return default data in case of an error
+            return new CritOverhaulData();
         }
     }
 
-    private void saveConfig(CritOverhaulData data) {
+    private void saveConfig(CritOverhaulData newData) {
         try {
             Path configFilePath = Paths.get(CONFIG_FILE_PATH);
             Files.createDirectories(configFilePath.getParent());
 
-            try (Writer writer = new FileWriter(configFilePath.toFile())) {
-                gson.toJson(data, writer);
+            if (!Files.exists(configFilePath) || Files.size(configFilePath) == 0) {
+                Files.createFile(configFilePath);
+                try (Writer writer = new FileWriter(configFilePath.toFile())) {
+                    gson.toJson(newData, writer);
+                }
             }
         } catch (IOException e) {
             e.printStackTrace();
         }
-    }
-
-    public CritOverhaulData getCritChanceData() {
-        return critOverhaulData;
     }
     public CritOverhaulEntry getCritDataForItem(String itemName) {
         return critOverhaulData.getItems().getOrDefault(itemName, new CritOverhaulEntry(0.0f, 0.0f));
